bgpd: resolve ipv6 ecmp issue with vrfs and ll nexthop
Problem reported that ecmp wasn't working correctly in a vrf with
ipv6.  Issue was that originator of the routes were sending the updates
with a link-local nexthop and nhlen of 16.  In this particular case,
bgp_zebra_announce was using the wrong call to get the ifindex and
was not supplying the vrf.   This caused ecmp to work only in the case
of the default vrf.

Ticket: CM-15545
Signed-off-by: Don Slice <dslice@cumulusnetworks.com>
Reviewed-by: CCR-6017