zebra: Distributed Dataplane module
Using the recently added feature to allow loading modules dynamically
with dlopen, a few hooks are added to be able to have a remote
dataplane. In particular, control is outsourced for:

- Add and delete IP addresses.
- Add and delete IP routes.
- Add and delete MPLS routes.
- Add and delete Pseudowires.

A new command line parameter, --dp-opts, is added to be able to tweak
module execution. It's up to the module to parse and use content passed
into it.

Besides, support for ZMQ communication is added to FRR thread library.

Signed-off-by: ßingen <bingen@voltanet.io>