lib: allow pthreads to poke select/poll
When scheduling a task onto a thread master owned by another pthread, we
need to lock the thread master's mutex. However, if the pthread which
owns that thread master is in poll() or select(), we could be stuck
waiting for a very long time.

To solve this problem, this change adds a pipe to struct thread_master.
The reading end is passed into select() or poll(). When another pthread
wants to lock the mutex but can't, it will write a byte into the pipe
and try again. The thread master which owns the pipe will flush out the
pipe so it can go back to sleep while waiting on I/O.

Have tested this strategy with 1000+ BGP peers and did not see serious
performance degradation, so I would say this scales well.

Also has a one-off fix for a null deref.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>