zebra: Distributed Dataplane module
Using the recently added feature to allow loading modules dynamically
with dlopen, a few hooks are added to be able to have a remote
dataplane. In particular, control is outsourced for:

- Add and delete IP addresses.
- Add and delete IP routes.
- Add and delete MPLS routes.

Once Zebra PW management is merge, it will be added too:

- Add and delete Pseudowires.

If an external module is enabled, installation of routes is held in
Zebra and the module is responsible for it.

A new command line parameter, --dp-opts, is added to be able to tweak
module execution. It's up to the module to parse and use content passed
into it.

Signed-off-by: ßingen <bingen@voltanet.io>