pimd: Add a check for needing spt bit in register_recv
When we receive a register message for an existing S,G
and the SPT bit is not set, then do a quick check against
the S,G to see if we have counters updated.  This is added
because the existing S,G wheel_timer only updates every
31 seconds and it is possible for a Null Register to
have come in before that first 30 second timer pops.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>